requirements.txt
pillow==7.2.0
aiohttp==3.6.2
aiogram==2.10.1